requirements.txt
docopt==0.6.2
elasticsearch==2.3.0
PyYAML==3.11
